project/doctor_service/requirements.txt
fastapi
uvicorn
pymongo[srv]
certifi
passlib[bcrypt]
bcrypt
python-jose[cryptography]
pydantic
python-dotenv
cryptography
